fix authorized all api

+++ src/controller/project.controller.js
@@ -47,22 +47,22 @@ exports.count = async (req, res) => {
 //[PUT] api/project/status/:id
 exports.updateStatus = async (req, res) => {
     try {
-        // if (req.role === "admin") {
-        const id = req.params.id;
-        const project = await Project.findOne({ _id: id });
-        project.status = project.status === "done" ? "undone" : "done";
-        await project.save();
-        res.status(200).json({
-            code: "200",
-            message: "Update status thành công!",
-            body: project,
-        });
-        // } else {
-        //     res.status(200).json({
-        //         code: "400",
-        //         message: "Bạn không có quyền update status project",
-        //     });
-        // }
+        if (req.role === "admin") {
+            const id = req.params.id;
+            const project = await Project.findOne({ _id: id });
+            project.status = project.status === "done" ? "undone" : "done";
+            await project.save();
+            res.status(200).json({
+                code: "200",
+                message: "Update status thành công!",
+                body: project,
+            });
+        } else {
+            res.status(200).json({
+                code: "400",
+                message: "Bạn không có quyền update status project",
+            });
+        }
     } catch (err) {
         return res
             .status(500)
@@ -73,23 +73,26 @@ exports.updateStatus = async (req, res) => {
 //[put] api/project/:id
 exports.update = async (req, res) => {
     try {
-        // if (req.role === "admin") {
-        const id = req.params.id;
-        const project = await Project.findOne({ _id: id });
-        if (project.length === 0 || typeof project === "undefined") {
-            return res.status(200).json({ message: "Project không tồn tại" });
+        if (req.role === "admin") {
+            const id = req.params.id;
+            const project = await Project.findOne({ _id: id });
+            if (project.length === 0 || typeof project === "undefined") {
+                return res
+                    .status(200)
+                    .json({ message: "Project không tồn tại" });
+            }
+            const projectUpdate = await projectService.update(req);
+            res.status(200).json({
+                code: "200",
+                message: "Update Project thành công!",
+                projectUpdate,
+            });
+        } else {
+            res.status(200).json({
+                code: "400",
+                message: "Bạn không có quyền sửa project",
+            });
         }
-        const projectUpdate = await projectService.update(req);
-        res.status(200).json({
-            code: "200",
-            message: "Update Project thành công!",
-            projectUpdate,
-        });
-        // } else {
-        //     res.status(200).json({
-        //         message: "Bạn không có quyền sửa thông tin project",
-        //     });
-        // }
     } catch (err) {
         return res.status(500).json("Server error " + err.message);
     }
@@ -97,27 +100,27 @@ exports.update = async (req, res) => {
 //[DELETE] /api/project/:id
 exports.delete = async (req, res) => {
     try {
-        // if (req.role === "admin") {
-        const id = req.params.id;
-        const project = await Project.findOne({ _id: id });
-        await Task.deleteMany({ project: id });
-        if (project.length === 0 || typeof project === "undefined") {
-            return res
-                .status(200)
-                .json({ code: "400", message: "Project không tồn tại" });
+        if (req.role === "admin") {
+            const id = req.params.id;
+            const project = await Project.findOne({ _id: id });
+            await Task.deleteMany({ project: id });
+            if (project.length === 0 || typeof project === "undefined") {
+                return res
+                    .status(200)
+                    .json({ code: "400", message: "Project không tồn tại" });
+            }
+            await Project.deleteOne({ _id: id });
+            res.status(200).json({
+                code: "200",
+                message: "Xóa project thành công!!",
+                body: project,
+            });
+        } else {
+            res.status(200).json({
+                code: "400",
+                message: "Bạn không có quyền xóa project",
+            });
         }
-        await Project.deleteOne({ _id: id });
-        res.status(200).json({
-            code: "200",
-            message: "Xóa project thành công!!",
-            body: project,
-        });
-        // } else {
-        //     res.status(200).json({
-        //         code: "400",
-        //         message: "Bạn không có quyền xóa project",
-        //     });
-        // }
     } catch (err) {
         return res
             .status(500)
@@ -127,30 +130,30 @@ exports.delete = async (req, res) => {
 //[POST] /api/project
 exports.create = async (req, res) => {
     try {
-        // if (req.role === "admin") {
-        const errors = validationResult(req);
-        if (!errors.isEmpty()) {
-            res.status(200).json({ errors: errors.array() });
-            return;
-        }
-        const project = await projectService.createService(req);
-        if (project.length === 0 || typeof project === "undefined") {
+        if (req.role === "admin") {
+            const errors = validationResult(req);
+            if (!errors.isEmpty()) {
+                res.status(200).json({ errors: errors.array() });
+                return;
+            }
+            const project = await projectService.createService(req);
+            if (project.length === 0 || typeof project === "undefined") {
+                res.status(200).json({
+                    code: "400",
+                    message: "Thêm project không thành công",
+                });
+            }
+            res.status(200).json({
+                code: "200",
+                message: "Thêm project thành công",
+                project: project,
+            });
+        } else {
             res.status(200).json({
                 code: "400",
-                message: "Thêm project không thành công",
+                message: "Bạn không có quyền thêm project",
             });
         }
-        res.status(200).json({
-            code: "200",
-            message: "Thêm project thành công",
-            project: project,
-        });
-        // } else {
-        //     res.status(200).json({
-        //         code: "400",
-        //         message: "Bạn không có quyền thêm project",
-        //     });
-        // }
     } catch (err) {
         // log.error(`Create project error: ${err.message}`);
         return res.status(500).json("Server error " + err.message);
@@ -194,39 +197,46 @@ exports.show = async (req, res) => {
 // [POST] api/project/user
 exports.addUser = async (req, res) => {
     try {
-        const idProject = req.params.idProject;
-        const idUser = req.params.idUser;
-        const user = await User.findOne({ _id: idUser });
-        const project = await Project.findOne({ _id: idProject });
-        if (project.length === 0 || typeof project === "undefined") {
-            throw Error("Not found project");
-        }
-        if (user === 0 || typeof user === "undefined") {
-            throw Error("Not found user");
-        }
-        if (!project.users.includes(idUser)) {
-            project.users.push(idUser);
-            await project.save();
-        } else {
+        if (req.role === "admin") {
+            const idProject = req.params.idProject;
+            const idUser = req.params.idUser;
+            const user = await User.findOne({ _id: idUser });
+            const project = await Project.findOne({ _id: idProject });
+            if (project.length === 0 || typeof project === "undefined") {
+                throw Error("Not found project");
+            }
+            if (user === 0 || typeof user === "undefined") {
+                throw Error("Not found user");
+            }
+            if (!project.users.includes(idUser)) {
+                project.users.push(idUser);
+                await project.save();
+            } else {
+                return res.status(200).json({
+                    code: "400",
+                    message: "Đã tồn tại user trong project này!!",
+                });
+            }
+            if (!user.projects.includes(idProject)) {
+                user.projects.push(idProject);
+                await user.save();
+            } else {
+                return res.status(200).json({
+                    code: "400",
+                    message: "Đã tồn tại user trong project này!!",
+                });
+            }
             return res.status(200).json({
+                code: "200",
+                message: "Thêm user vào project thành công!!",
+                project,
+            });
+        } else {
+            res.status(200).json({
                 code: "400",
-                message: "Đã tồn tại user trong project này!!",
+                message: "Bạn không có quyền thêm user",
             });
         }
-        if (!user.projects.includes(idProject)) {
-            user.projects.push(idProject);
-            await user.save();
-        } else {
-            return res.status(200).json({
-                code: "400",
-                message: "Đã tồn tại user trong project này!!",
-            });
-        }
-        return res.status(200).json({
-            code: "200",
-            message: "Thêm user vào project thành công!!",
-            project,
-        });
     } catch (err) {
         return res.status(500).json({ code: "500", message: err.message });
     }
